refactor(settings): split processor, shared scales/feed, trim allowlist
- Remove dead VideoSection; scales + feeder relay in shared components
- ProcessorSection -> eight blocks under sections/processor/
- AUTO_ALLOWLIST_KEYS: drop keys already bound in Settings form fields
- Legacy integrations.scales.motion_trigger_* stay explicit allowlist
- CHANGELOG [Unreleased]

Made-with: Cursor
