+++ app/ui/src/pages/Settings/sections/CaptureFeederSection.tsx
@@ -18,13 +18,14 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { ServiceBlock } from '../shared/ServiceBlock';
+import { ScalesIntegrationFields } from '../shared/scalesIntegrationFields';
+import { FeederRelayFields } from '../shared/feederRelayFields';
 import type { Settings } from '../../../types';
 
 type Props = {
   form: ReactFormExtendedApi<Settings, undefined>;
 };
 
-type ScalesSource = NonNullable<NonNullable<Settings['integrations']>['scales']>['source'];
 type TriggerTransportSource = NonNullable<
   NonNullable<Settings['triggers']>['motion_sensor']
 >['source'];
@@ -358,234 +359,7 @@ export function CaptureFeederSection({ form }: Props) {
                   <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
                     {t('settings.serviceScalesDesc')}
                   </Typography>
-                  <Grid container spacing={2}>
-                    <Grid size={{ xs: 12 }}>
-                      <form.Field name="integrations.scales.enabled">
-                        {(field) => (
-                          <FormControlLabel
-                            control={
-                              <Switch
-                                checked={field.state.value ?? false}
-                                onChange={(e) => field.handleChange(e.target.checked)}
-                              />
-                            }
-                            label={t('settings.scalesEnabled')}
-                          />
-                        )}
-                      </form.Field>
-                    </Grid>
-                    <form.Subscribe selector={(s) => s.values.integrations?.scales?.enabled}>
-                      {(scalesOn) =>
-                        scalesOn ? (
-                          <>
-                            <Grid size={{ xs: 12 }}>
-                              <form.Field name="integrations.scales.source">
-                                {(field) => (
-                                  <FormControl fullWidth>
-                                    <InputLabel id="settings-scales-src">{t('settings.scalesSource')}</InputLabel>
-                                    <Select
-                                      labelId="settings-scales-src"
-                                      value={field.state.value ?? 'mqtt'}
-                                      label={t('settings.scalesSource')}
-                                      onChange={(e) => field.handleChange(e.target.value as ScalesSource)}
-                                    >
-                                      <MenuItem value="mqtt">{t('settings.scalesSourceMqtt')}</MenuItem>
-                                      <MenuItem value="esphome">{t('settings.scalesSourceEsp')}</MenuItem>
-                                    </Select>
-                                    <FormHelperText>{t('settings.scalesSourceHint')}</FormHelperText>
-                                  </FormControl>
-                                )}
-                              </form.Field>
-                            </Grid>
-                            <form.Subscribe selector={(s) => s.values.integrations?.scales?.source}>
-                              {(src) => (
-                                <>
-                                  {(src ?? 'mqtt') === 'mqtt' ? (
-                                    <>
-                                      <Grid size={{ xs: 12 }}>
-                                        <Alert severity="info">{t('settings.scalesMqttAlert')}</Alert>
-                                      </Grid>
-                                      <Grid size={{ xs: 12 }}>
-                                        <form.Field name="integrations.scales.mqtt_topic_prefix">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesMqttPrefix')}
-                                              placeholder="birdlense/scale"
-                                              helperText={t('settings.scalesMqttPrefixHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12 }}>
-                                        <form.Field name="integrations.scales.mqtt_topic">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesMqttTopic')}
-                                              placeholder="birdlense/scale/weight"
-                                              helperText={t('settings.scalesMqttTopicHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12 }}>
-                                        <form.Field name="integrations.scales.mqtt_bird_present_topic">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesMqttBirdPresentTopic')}
-                                              placeholder="birdlense/scale/bird_present"
-                                              helperText={t('settings.scalesMqttBirdPresentTopicHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12 }}>
-                                        <form.Field name="integrations.scales.mqtt_command_topic">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesMqttCommandTopic')}
-                                              placeholder="birdlense/scale/command"
-                                              helperText={t('settings.scalesMqttCommandTopicHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                    </>
-                                  ) : (
-                                    <>
-                                      <Grid size={{ xs: 12 }}>
-                                        <Alert severity="info">{t('settings.scalesEspAlert')}</Alert>
-                                      </Grid>
-                                      <Grid size={{ xs: 12, sm: 6 }}>
-                                        <form.Field name="integrations.scales.esphome_url">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.esphomeUrl')}
-                                              placeholder="http://192.168.1.50"
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12, sm: 6 }}>
-                                        <form.Field name="integrations.scales.esphome_weight_sensor_id">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesEspWeightSensorId')}
-                                              placeholder="weight_live_internal"
-                                              helperText={t('settings.scalesEspWeightSensorIdHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12, sm: 6 }}>
-                                        <form.Field name="integrations.scales.esphome_bird_present_sensor_id">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesEspBirdSensorId')}
-                                              placeholder="bird_present"
-                                              helperText={t('settings.scalesEspBirdSensorIdHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <Grid size={{ xs: 12, sm: 6 }}>
-                                        <form.Field name="integrations.scales.esphome_tare_button_id">
-                                          {(field) => (
-                                            <TextField
-                                              fullWidth
-                                              value={field.state.value ?? ''}
-                                              onChange={(e) => field.handleChange(e.target.value)}
-                                              label={t('settings.scalesEspTareButtonId')}
-                                              placeholder="manual_tare"
-                                              helperText={t('settings.scalesEspTareButtonIdHint')}
-                                            />
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                    </>
-                                  )}
-                                  {(src ?? 'mqtt') === 'mqtt' ? (
-                                    <>
-                                      <Grid size={{ xs: 12 }}>
-                                        <form.Field name="integrations.scales.weight_estimate_enabled">
-                                          {(field) => (
-                                            <>
-                                              <FormControlLabel
-                                                control={
-                                                  <Switch
-                                                    checked={field.state.value ?? true}
-                                                    onChange={(e) => field.handleChange(e.target.checked)}
-                                                  />
-                                                }
-                                                label={t('settings.scalesWeightEstimate')}
-                                              />
-                                              <Typography variant="body2" color="text.secondary" display="block">
-                                                {t('settings.scalesWeightEstimateHint')}
-                                              </Typography>
-                                            </>
-                                          )}
-                                        </form.Field>
-                                      </Grid>
-                                      <form.Subscribe
-                                        selector={(s) =>
-                                          s.values.integrations?.scales?.weight_estimate_enabled !== false
-                                        }
-                                      >
-                                        {(weOn) =>
-                                          weOn ? (
-                                            <Grid size={{ xs: 12 }}>
-                                              <form.Field name="integrations.scales.estimate_require_consecutive_spike">
-                                                {(field) => (
-                                                  <>
-                                                    <FormControlLabel
-                                                      control={
-                                                        <Switch
-                                                          checked={field.state.value ?? true}
-                                                          onChange={(e) => field.handleChange(e.target.checked)}
-                                                        />
-                                                      }
-                                                      label={t('settings.scalesEstimateRequireSpike')}
-                                                    />
-                                                    <Typography variant="body2" color="text.secondary" display="block">
-                                                      {t('settings.scalesEstimateRequireSpikeHint')}
-                                                    </Typography>
-                                                  </>
-                                                )}
-                                              </form.Field>
-                                            </Grid>
-                                          ) : null
-                                        }
-                                      </form.Subscribe>
-                                    </>
-                                  ) : null}
-                                </>
-                              )}
-                            </form.Subscribe>
-                          </>
-                        ) : null
-                      }
-                    </form.Subscribe>
-                  </Grid>
+                  <ScalesIntegrationFields form={form} />
                 </ServiceBlock>
               </Grid>
 
@@ -698,140 +472,7 @@ export function CaptureFeederSection({ form }: Props) {
             </Grid>
           </ServiceBlock>
 
-          <ServiceBlock title={t('settings.accordionFeed')}>
-            <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
-              {t('settings.accordionFeedDesc')}
-            </Typography>
-            <Grid container spacing={2}>
-              <Grid size={{ xs: 12 }}>
-                <form.Field name="feed.source">
-                  {(field) => (
-                    <FormControl fullWidth>
-                      <InputLabel id="settings-feed-type-label">{t('settings.feedType')}</InputLabel>
-                      <Select
-                        labelId="settings-feed-type-label"
-                        value={field.state.value ?? 'none'}
-                        label={t('settings.feedType')}
-                        onChange={(e) => field.handleChange(e.target.value)}
-                      >
-                        <MenuItem value="none">{t('settings.feedNone')}</MenuItem>
-                        <MenuItem value="mqtt">{t('settings.feedMqtt')}</MenuItem>
-                        <MenuItem value="esphome">{t('settings.feedEsp')}</MenuItem>
-                      </Select>
-                    </FormControl>
-                  )}
-                </form.Field>
-              </Grid>
-              <form.Subscribe selector={(state) => state.values.feed?.source}>
-                {(source) => (
-                  <>
-                    {source === 'mqtt' && (
-                      <>
-                        <Grid size={{ xs: 12 }}>
-                          <Alert severity="info" sx={{ mb: 2 }}>
-                            {t('settings.feedMqttHint')}
-                          </Alert>
-                        </Grid>
-                        <Grid size={{ xs: 12, sm: 6 }}>
-                          <form.Field name="feed.mqtt_topic">
-                            {(field) => (
-                              <TextField
-                                fullWidth
-                                value={field.state.value ?? ''}
-                                onChange={(e) => field.handleChange(e.target.value)}
-                                label={t('settings.relayTopic')}
-                                placeholder="cmnd/bird_feeder/Power"
-                                helperText={t('settings.relayTopicHint')}
-                              />
-                            )}
-                          </form.Field>
-                        </Grid>
-                      </>
-                    )}
-                    {source === 'esphome' && (
-                      <>
-                        <Grid size={{ xs: 12 }}>
-                          <Alert severity="info" sx={{ mb: 2 }}>
-                            {t('settings.esphomeFeedHint')}
-                          </Alert>
-                        </Grid>
-                        <Grid size={{ xs: 12, sm: 6 }}>
-                          <form.Field name="feed.esphome_type">
-                            {(field) => (
-                              <FormControl fullWidth>
-                                <InputLabel id="settings-switch-type-label">
-                                  {t('settings.switchType')}
-                                </InputLabel>
-                                <Select
-                                  labelId="settings-switch-type-label"
-                                  value={field.state.value ?? 'switch'}
-                                  label={t('settings.switchType')}
-                                  onChange={(e) =>
-                                    field.handleChange(e.target.value as 'switch' | 'button')
-                                  }
-                                >
-                                  <MenuItem value="switch">
-                                    {t('settings.switchTypeSwitch')}
-                                  </MenuItem>
-                                  <MenuItem value="button">
-                                    {t('settings.switchTypeButton')}
-                                  </MenuItem>
-                                </Select>
-                              </FormControl>
-                            )}
-                          </form.Field>
-                        </Grid>
-                        <Grid size={{ xs: 12, sm: 6 }}>
-                          <form.Field name="feed.esphome_url">
-                            {(field) => (
-                              <TextField
-                                fullWidth
-                                value={field.state.value ?? ''}
-                                onChange={(e) => field.handleChange(e.target.value)}
-                                label={t('settings.deviceUrl')}
-                                placeholder="http://192.168.1.50"
-                              />
-                            )}
-                          </form.Field>
-                        </Grid>
-                        <Grid size={{ xs: 12, sm: 6 }}>
-                          <form.Field name="feed.esphome_switch_id">
-                            {(field) => (
-                              <TextField
-                                fullWidth
-                                value={field.state.value ?? ''}
-                                onChange={(e) => field.handleChange(e.target.value)}
-                                label={t('settings.switchId')}
-                                placeholder="bird_feeder"
-                                helperText={t('settings.switchIdHint')}
-                              />
-                            )}
-                          </form.Field>
-                        </Grid>
-                      </>
-                    )}
-                    {(source === 'mqtt' || source === 'esphome') && (
-                      <Grid size={{ xs: 12, sm: 6 }}>
-                        <form.Field name="feed.duration_seconds">
-                          {(field) => (
-                            <TextField
-                              fullWidth
-                              type="number"
-                              inputProps={{ min: 1, max: 30 }}
-                              value={field.state.value ?? 3}
-                              onChange={(e) => field.handleChange(Number(e.target.value) || 3)}
-                              label={t('settings.relaySeconds')}
-                              helperText={t('settings.relaySecondsHint')}
-                            />
-                          )}
-                        </form.Field>
-                      </Grid>
-                    )}
-                  </>
-                )}
-              </form.Subscribe>
-            </Grid>
-          </ServiceBlock>
+          <FeederRelayFields form={form} />
 
         </Box>
       </AccordionDetails>
