+++ app/ui/src/pages/Settings/sections/processor/ProcessorLightGateBlock.tsx
@@ -1,0 +1,83 @@
+import { useTranslation } from 'react-i18next';
+import type { ReactFormExtendedApi } from '@tanstack/react-form';
+import Grid from '@mui/material/Grid2';
+import TextField from '@mui/material/TextField';
+import Typography from '@mui/material/Typography';
+import Checkbox from '@mui/material/Checkbox';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import FormControl from '@mui/material/FormControl';
+import FormHelperText from '@mui/material/FormHelperText';
+import { ServiceBlock } from '../../shared/ServiceBlock';
+import type { Settings } from '../../../../types';
+
+type Props = {
+  form: ReactFormExtendedApi<Settings, undefined>;
+};
+
+export function ProcessorLightGateBlock({ form }: Props) {
+  const { t } = useTranslation();
+
+  return (
+    <ServiceBlock title={t('settings.lightGateTitle')}>
+      <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
+        {t('settings.lightGateDesc')}
+      </Typography>
+      <Grid container spacing={2}>
+        <Grid size={{ xs: 12 }}>
+          <form.Field name="processor.light_gate_enabled">
+            {(field) => (
+              <FormControl fullWidth>
+                <FormControlLabel
+                  control={
+                    <Checkbox
+                      checked={field.state.value !== false}
+                      onChange={(e) => field.handleChange(e.target.checked)}
+                    />
+                  }
+                  label={t('settings.lightGateEnabled')}
+                />
+                <FormHelperText>
+                  {t('settings.lightGateEnabledHelp')}
+                </FormHelperText>
+              </FormControl>
+            )}
+          </form.Field>
+        </Grid>
+        <Grid size={{ xs: 12, sm: 6 }}>
+          <form.Field name="processor.light_gate_min_brightness">
+            {(field) => (
+              <TextField
+                fullWidth
+                type="number"
+                inputProps={{ min: 0, max: 255, step: 1 }}
+                value={field.state.value ?? 25}
+                onChange={(e) =>
+                  field.handleChange(Number(e.target.value) || undefined)
+                }
+                label={t('settings.lightGateMinBrightness')}
+                helperText={t('settings.lightGateMinBrightnessHelp')}
+              />
+            )}
+          </form.Field>
+        </Grid>
+        <Grid size={{ xs: 12, sm: 6 }}>
+          <form.Field name="processor.light_gate_min_contrast">
+            {(field) => (
+              <TextField
+                fullWidth
+                type="number"
+                inputProps={{ min: 0, max: 255, step: 1 }}
+                value={field.state.value ?? 20}
+                onChange={(e) =>
+                  field.handleChange(Number(e.target.value) || undefined)
+                }
+                label={t('settings.lightGateMinContrast')}
+                helperText={t('settings.lightGateMinContrastHelp')}
+              />
+            )}
+          </form.Field>
+        </Grid>
+      </Grid>
+    </ServiceBlock>
+  );
+}
