+++ scripts/check-settings-ui-coverage.py
@@ -33,6 +33,21 @@ ALLOWED_NON_UI_KEYS: dict[str, dict[str, str]] = {
         "reason": "Processor resolves scales trigger transport from integrations.scales.source when unset; duplicate source pickers were removed from Settings.",
         "next_step": "Re-expose only if product needs different MQTT/ESPHome paths for live weight vs weight-trigger sampling.",
     },
+    "integrations.scales.motion_trigger_enabled": {
+        "category": "legacy",
+        "reason": "Superseded by triggers.scales.enabled; trigger_config still reads this for YAML migration.",
+        "next_step": "Remove key after migration period or document YAML-only.",
+    },
+    "integrations.scales.motion_trigger_min_delta_kg": {
+        "category": "legacy",
+        "reason": "Superseded by triggers.scales.motion_trigger_min_delta_kg with fallback in trigger_config.",
+        "next_step": "YAML-only once old configs are rare.",
+    },
+    "integrations.scales.motion_trigger_debounce_seconds": {
+        "category": "legacy",
+        "reason": "Superseded by triggers.scales.motion_trigger_debounce_seconds with fallback in trigger_config.",
+        "next_step": "YAML-only once old configs are rare.",
+    },
     # Processor internals.
     "processor.detection_strategy": {
         "category": "advanced",
@@ -215,152 +230,53 @@ AUTO_ALLOWLIST_KEYS: set[str] = {
     "detection.fusion_alpha",
     "detection.fusion_model_path",
     "detection.use_learned_fusion",
-    "ebird.country",
-    "ebird.location_name",
-    "ebird.species_mapping",
-    "ebird.state",
-    "feed.duration_seconds",
-    "feed.esphome_switch_id",
-    "feed.esphome_type",
-    "feed.esphome_url",
-    "feed.mqtt_topic",
-    "feed.source",
-    "general.birdnet_url",
-    "general.contributor_password",
-    "general.donate_url",
-    "general.enable_notifications",
-    "general.notification_excluded_species",
     "general.require_auth_for_video_stream",
-    "general.settings_password",
-    "integrations.scales.enabled",
     "integrations.scales.history_max_lines",
-    "integrations.scales.esphome_bird_present_sensor_id",
-    "integrations.scales.esphome_tare_button_id",
-    "integrations.scales.esphome_url",
-    "integrations.scales.esphome_weight_sensor_id",
     "integrations.scales.homeassistant_entity_id",
     "integrations.scales.min_delta_kg_for_estimate",
-    "integrations.scales.mqtt_bird_present_topic",
-    "integrations.scales.mqtt_command_topic",
     "integrations.scales.mqtt_tare_payload",
-    "integrations.scales.mqtt_topic",
-    "integrations.scales.source",
     "integrations.scales.unit",
-    "mcp.enabled",
-    "mcp.token",
     "motion.check_every_n_frames",
-    "motion.frigate_min_trigger_score",
     "motion.esphome_sensor_id",
     "motion.esphome_url",
+    "motion.frigate_min_trigger_score",
+    "motion.mqtt_topic",
     "motion.opencv_diff_threshold",
     "motion.opencv_min_contour_area",
-    "motion.mqtt_topic",
     "motion.source",
-    "mqtt.birdnet_topic",
-    "mqtt.broker",
-    "mqtt.frigate_topic",
-    "mqtt.ha_discovery",
-    "mqtt.port",
-    "mqtt.publish_topic",
-    "mqtt.reconnect_max_delay",
-    "mqtt.reconnect_min_delay",
-    "notifications.base_url",
-    "notifications.compress_photo_over_kb",
-    "notifications.custom_emoji_id_bird",
-    "notifications.custom_emoji_id_chipmunk",
-    "notifications.custom_emoji_id_open_live",
-    "notifications.disable_notification",
-    "notifications.link_preview_large",
-    "notifications.message_thread_id",
-    "notifications.paid_media_forward_star_count",
-    "notifications.paid_media_view_star_count",
-    "notifications.protect_content",
-    "notifications.send_photo",
-    "notifications.telegram_api_base",
-    "notifications.telegram_api_hash",
-    "notifications.telegram_api_id",
-    "notifications.telegram_bot_token",
-    "notifications.telegram_chat_id",
-    "notifications.telegram_max_side_px",
-    "notifications.telegram_mtproto_host",
-    "notifications.telegram_mtproto_port",
-    "notifications.telegram_mtproto_secret",
-    "notifications.telegram_proxy_type",
-    "notifications.telegram_proxy_url",
-    "notifications.telegram_retries",
-    "notifications.telegram_timeout",
-    "notifications.use_custom_emoji",
-    "performance.cache_redis_enabled",
-    "performance.redis_url",
-    "processor.birdnet_mqtt_auto_confidence",
-    "processor.birdnet_mqtt_bias_delta",
     "processor.binary_imgsz",
+    "processor.birdnet_mqtt_bias_window_seconds",
     "processor.birdnet_mqtt_prior_half_life_hours",
     "processor.birdnet_mqtt_prior_min_confidence",
     "processor.birdnet_mqtt_prior_ttl_hours",
     "processor.birdnet_mqtt_prior_window_hours",
-    "processor.birdnet_mqtt_bias_window_seconds",
     "processor.blur_threshold",
     "processor.classification_scheduler",
-    "processor.key_frame_limit",
-    "processor.max_blur_checks",
-    "processor.max_classifications_per_frame",
-    "processor.min_center_dist",
-    "processor.birdnet_mqtt_bias_floor",
-    "processor.dataset_min_confidence",
-    "processor.ebird_regional_top_auto_confidence",
-    "processor.ebird_regional_top_confidence_delta",
-    "processor.ebird_regional_top_confidence_floor",
-    "processor.inference_lores_px",
-    "processor.max_inactive_seconds",
-    "processor.max_record_seconds",
     "processor.frigate_activity_hold_seconds",
     "processor.generic_bird_min_area_frac",
     "processor.generic_bird_min_best_frame_score",
     "processor.generic_bird_min_detector_conf",
     "processor.generic_bird_min_frames",
+    "processor.inference_lores_px",
+    "processor.key_frame_limit",
+    "processor.max_blur_checks",
+    "processor.max_classifications_per_frame",
+    "processor.min_center_dist",
     "processor.min_seconds_between_recordings",
-    "processor.min_confidence_binary",
-    "processor.min_confidence_to_notify",
-    "processor.min_confidence_to_process",
-    "processor.min_track_duration",
-    "processor.multi_camera_confidence_boost",
-    "processor.multi_camera_groups",
-    "processor.post_record_seconds",
-    "processor.save_dataset_crops",
-    "processor.spectrogram_px_per_sec",
-    "processor.generate_spectrogram_always",
-    "processor.tracker",
     "processor.track_regen_detection_strategy",
     "processor.track_regen_frame_step",
     "processor.track_regen_ignore_regional_species",
-    "processor.track_regen_match_live_pipeline",
     "processor.track_regen_lores_px",
+    "processor.track_regen_match_live_pipeline",
     "processor.track_regen_precise_detection_strategy",
     "processor.track_regen_precise_min_center_dist",
     "processor.track_regen_precise_timeout_sec",
     "processor.track_regen_video_timeout_sec",
-    "secrets.ebird_api_key",
-    "secrets.latitude",
-    "secrets.longitude",
-    "secrets.openweather_api_key",
-    "secrets.xeno_canto_api_key",
     "species.catalog_allowlist_file",
     "species.catalog_strict_ingest",
     "species.tuning_target_species_ids",
-    "ui.unknown_confidence_threshold",
-    "video.cameras",
-    "video.encoding",
     "video.file_path",
     "video.file_realtime_simulation",
-    "video.go2rtc_url",
-    "video.video_height",
-    "video.video_width",
-    "homeassistant.token",
-    "homeassistant.url",
-    "weather.ha_entity_id",
-    "weather.source",
-    "webhook.url",
 }
 
 TERMINAL_MAP_KEYS = {
